refactor: 채팅 컴포넌트 패키지명 변경 #140
