+++ src/components/Chatting/index.ts
@@ -1,5 +1,5 @@
-export { default as ChatContent } from './Content';
-export { default as ChatMain } from './Main';
+export { default as ChatContent } from './ChatListContent';
+export { default as ChatMain } from './ChatListWrapper';
 export { default as ChatRoomHeader } from './ChatRoomHeader';
 export { default as ChatRoomContent } from './ChatRoomContent';
 export { default as ChatRoomBottom } from './ChatRoomBottom';
